[WDP][Search] Add findBestMatch & other small interface changes
This adds the function findBestMatch to the Search module which searches
through a second image to find a neighbourhood which has the lowest sum
of squared differences to a neighbourhood in the first image.

In addition the wording of 'block' has been changed to 'neighbourhood'
throughout, code has been made more readable, and some assertions were
added to catch out of bounds errors early.